+++ packages/pangea-sdk/PangeaCyber.Net/AIGuard/Requests/TextGuardRequest.cs
@@ -23,8 +23,8 @@ public sealed class TextGuardRequest : BaseRequest
     /// <summary>Setting this value to true will provide a detailed analysis of the text data</summary>
     public bool Debug { get; set; }
 
-    /// <summary>Short string hint for the LLM Provider information</summary>
-    public string? LlmInfo { get; set; }
+    /// <summary>Overrides.</summary>
+    public Overrides? Overrides { get; set; }
 
     /// <summary>Additional fields to include in activity log</summary>
     public LogFields? LogFields { get; set; }