+++ packages/pangea-sdk/PangeaCyber.Tests/ITRedactTest.cs
@@ -1,13 +1,13 @@
 using PangeaCyber.Net;
-using PangeaCyber.Net.Redact;
 using PangeaCyber.Net.Exceptions;
+using PangeaCyber.Net.Redact;
 
 namespace PangeaCyber.Tests;
 
 ///
 public class ITRedactTest
 {
-    private RedactClient client;
+    private readonly RedactClient client;
     private readonly TestEnvironment environment = Helper.LoadTestEnvironment("redact", TestEnvironment.LVE);
 
     ///
@@ -20,7 +20,7 @@ public class ITRedactTest
     [Fact]
     public async Task TestRedactText()
     {
-        var response = await client.RedactText(new RedactTextRequest.Builder("Jenny Jenny... 415-867-5309").WithReturnResult(true).Build());
+        var response = await client.RedactText(new RedactTextRequest("Jenny Jenny... 415-867-5309") { ReturnResult = true });
 
         Assert.True(response.IsOK);
         Assert.Equal("<PERSON>... <PHONE_NUMBER>", response.Result.RedactedText);
@@ -31,7 +31,7 @@ public class ITRedactTest
     [Fact]
     public async Task TestRedactText_2()
     {
-        var response = await client.RedactText(new RedactTextRequest.Builder("Jenny Jenny... 415-867-5309").WithReturnResult(true).WithDebug(true).Build());
+        var response = await client.RedactText(new RedactTextRequest("Jenny Jenny... 415-867-5309") { Debug = true, ReturnResult = true });
 
         Assert.True(response.IsOK);
         Assert.Equal("<PERSON>... <PHONE_NUMBER>", response.Result.RedactedText);
@@ -42,14 +42,13 @@ public class ITRedactTest
     [Fact]
     public async Task TestRedactText_3()
     {
-        var response = await client.RedactText(new RedactTextRequest.Builder("Jenny Jenny... 415-867-5309").WithReturnResult(false).WithDebug(true).Build());
+        var response = await client.RedactText(new RedactTextRequest("Jenny Jenny... 415-867-5309") { Debug = true, ReturnResult = false });
 
         Assert.True(response.IsOK);
         Assert.Null(response.Result.RedactedText);
         Assert.Equal(2, response.Result.Count);
         Assert.NotNull(response.Result.Report);
     }
-
 
     [Fact]
     public async Task TestRedactStructured()
@@ -61,7 +60,7 @@ public class ITRedactTest
             { "IP", "Its ip is 127.0.0.1"}
         };
 
-        var response = await client.RedactStructured(new RedactStructuredRequest.Builder(data).WithReturnResult(true).Build());
+        var response = await client.RedactStructured(new RedactStructuredRequest(data) { ReturnResult = true });
         var converted = ((Newtonsoft.Json.Linq.JObject)response.Result.RedactedData).ToObject<Dictionary<string, object>>();
 
         var expected = new Dictionary<string, object>
@@ -87,7 +86,7 @@ public class ITRedactTest
             { "IP", "Its ip is 127.0.0.1"}
         };
 
-        var response = await client.RedactStructured(new RedactStructuredRequest.Builder(data).WithFormat("json").WithReturnResult(true).Build());
+        var response = await client.RedactStructured(new RedactStructuredRequest(data) { Format = "json", ReturnResult = true });
         var converted = ((Newtonsoft.Json.Linq.JObject)response.Result.RedactedData).ToObject<Dictionary<string, object>>();
 
         var expected = new Dictionary<string, object>
@@ -112,7 +111,7 @@ public class ITRedactTest
             { "IP", "Its ip is 127.0.0.1"}
         };
 
-        var response = await client.RedactStructured(new RedactStructuredRequest.Builder(data).WithFormat("json").WithDebug(true).WithReturnResult(true).Build());
+        var response = await client.RedactStructured(new RedactStructuredRequest(data) { Debug = true, Format = "json", ReturnResult = true });
         var converted = ((Newtonsoft.Json.Linq.JObject)response.Result.RedactedData).ToObject<Dictionary<string, object>>();
 
         var expected = new Dictionary<string, object>
@@ -127,7 +126,6 @@ public class ITRedactTest
         Assert.NotNull(response.Result.Report);
     }
 
-
     [Fact]
     public async Task TestRedactStructured_4()
     {
@@ -138,7 +136,7 @@ public class ITRedactTest
             { "IP", "Its ip is 127.0.0.1"}
         };
 
-        var response = await client.RedactStructured(new RedactStructuredRequest.Builder(data).WithFormat("json").WithJsonp(new String[] { "Phone" }).WithReturnResult(true).Build());
+        var response = await client.RedactStructured(new RedactStructuredRequest(data) { Format = "json", Jsonp = ["Phone"], ReturnResult = true });
         var converted = ((Newtonsoft.Json.Linq.JObject)response.Result.RedactedData).ToObject<Dictionary<string, object>>();
 
         var expected = new Dictionary<string, object>
@@ -161,32 +159,32 @@ public class ITRedactTest
         {
             { "Name", "Jenny Jenny" },
             { "Phone", "This is its number: 415-867-5309" },
-            { "IP", "Its ip is 127.0.0.1"}
+            { "IP", "Its ip is 127.0.0.1" }
         };
 
-        var response = await client.RedactStructured(new RedactStructuredRequest.Builder(data).WithFormat("json").WithRules(new[] { "IP_ADDRESS" }).WithReturnResult(true).Build());
+        var response = await client.RedactStructured(new RedactStructuredRequest(data) { Format = "json", Rules = ["IP_ADDRESS"], ReturnResult = true });
         var converted = ((Newtonsoft.Json.Linq.JObject)response.Result.RedactedData).ToObject<Dictionary<string, object>>();
 
         var expected = new Dictionary<string, object>
         {
-            { "Name", "<PERSON>" },
-            { "Phone", "This is its number: <PHONE_NUMBER>" },
-            { "IP", "Its ip is <IP_ADDRESS>"}
+            { "Name", "Jenny Jenny" },
+            { "Phone", "This is its number: 415-867-5309" },
+            { "IP", "Its ip is <IP_ADDRESS>" }
         };
 
         Assert.True(response.IsOK);
         Assert.Equal(expected, converted);
-        Assert.Equal(3, response.Result.Count);
+        Assert.Equal(1, response.Result.Count);
         Assert.Null(response.Result.Report);
     }
 
     [Fact]
-    public async Task testRedactTextUnauthorized()
+    public async Task TestRedactTextUnauthorized()
     {
         Config cfg = Config.FromIntegrationEnvironment(environment);
         cfg.Token = "notarealtoken";
         var fakeClient = new RedactClient.Builder(cfg).Build();
-        await Assert.ThrowsAsync<UnauthorizedException>(async () => await fakeClient.RedactText(new RedactTextRequest.Builder("Jenny Jenny... 415-867-5309").WithReturnResult(true).Build()));
+        await Assert.ThrowsAsync<UnauthorizedException>(async () => await fakeClient.RedactText(new RedactTextRequest("Jenny Jenny... 415-867-5309") { ReturnResult = true }));
     }
 
     [Fact]
@@ -201,17 +199,17 @@ public class ITRedactTest
         Config cfg = Config.FromIntegrationEnvironment(environment);
         cfg.Token = "notarealtoken";
         var fakeClient = new RedactClient.Builder(cfg).Build();
-        await Assert.ThrowsAsync<UnauthorizedException>(async () => await fakeClient.RedactStructured(new RedactStructuredRequest.Builder(data).WithReturnResult(true).Build()));
+        await Assert.ThrowsAsync<UnauthorizedException>(async () => await fakeClient.RedactStructured(new RedactStructuredRequest(data) { ReturnResult = true }));
     }
 
     [Fact]
     public async Task TestRedactMultiConfig_1()
     {
         var cfg = new Config(Config.GetMultiConfigTestToken(environment), Config.GetTestDomain(environment));
-        String ConfigID = Config.GetConfigID(environment, RedactClient.ServiceName, 1);
+        var ConfigID = Config.GetConfigID(environment, RedactClient.ServiceName, 1);
         var clientMultiConfig = new RedactClient.Builder(cfg).WithConfigID(ConfigID).Build();
 
-        var response = await clientMultiConfig.RedactText(new RedactTextRequest.Builder("Jenny Jenny... 415-867-5309").WithReturnResult(true).Build());
+        var response = await clientMultiConfig.RedactText(new RedactTextRequest("Jenny Jenny... 415-867-5309") { ReturnResult = true });
 
         Assert.True(response.IsOK);
         Assert.Equal("<PERSON>... <PHONE_NUMBER>", response.Result.RedactedText);
@@ -223,10 +221,10 @@ public class ITRedactTest
     public async Task TestLogMultiConfig_2()
     {
         var cfg = new Config(Config.GetMultiConfigTestToken(environment), Config.GetTestDomain(environment));
-        String ConfigID = Config.GetConfigID(environment, RedactClient.ServiceName, 2);
+        var ConfigID = Config.GetConfigID(environment, RedactClient.ServiceName, 2);
         var clientMultiConfig = new RedactClient.Builder(cfg).WithConfigID(ConfigID).Build();
 
-        var response = await clientMultiConfig.RedactText(new RedactTextRequest.Builder("Jenny Jenny... 415-867-5309").WithReturnResult(true).Build());
+        var response = await clientMultiConfig.RedactText(new RedactTextRequest("Jenny Jenny... 415-867-5309") { ReturnResult = true });
 
         Assert.True(response.IsOK);
         Assert.Equal("<PERSON>... <PHONE_NUMBER>", response.Result.RedactedText);
@@ -240,6 +238,6 @@ public class ITRedactTest
         var cfg = new Config(Config.GetMultiConfigTestToken(environment), Config.GetTestDomain(environment));
         var clientMultiConfig = new RedactClient.Builder(cfg).Build();
 
-        await Assert.ThrowsAsync<PangeaAPIException>(async () => await clientMultiConfig.RedactText(new RedactTextRequest.Builder("Jenny Jenny... 415-867-5309").WithReturnResult(true).Build()));
+        await Assert.ThrowsAsync<PangeaAPIException>(async () => await clientMultiConfig.RedactText(new RedactTextRequest("Jenny Jenny... 415-867-5309") { ReturnResult = true }));
     }
 }